requirements.txt
# Core ML and Data Science
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
scipy>=1.9.0

# Natural Language Processing
nltk>=3.7
textblob>=0.17.1

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.10.0

# MLOps and Experiment Tracking
dvc>=2.40.0
mlflow>=2.0.0

# Configuration and Data Handling
pyyaml>=6.0
python-dotenv>=0.19.0
configparser>=5.2.0

# API and Deployment
flask>=2.2.0
gunicorn>=20.1.0
requests>=2.28.0
streamlit>=1.28.0

# Development and Testing
pytest>=7.0.0
pytest-cov>=4.0.0
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0
python-dateutil>=2.8.0

# Jupyter and Notebooks
jupyter>=1.0.0
ipykernel>=6.15.0
notebook>=6.4.0

# Security
safety>=2.3.0
bandit>=1.7.0

# Additional ML Libraries (optional)
xgboost>=1.6.0
lightgbm>=3.3.0

# Neural Networks (if using)
tensorflow>=2.10.0
torch>=1.12.0

# Text Processing
spacy>=3.4.0
wordcloud>=1.9.0